Split camera-backed Light tool modals
