+++ tests/test-correctness-phase2.js
@@ -164,8 +164,15 @@ const pwaAppShellAssets = [
   '/js/light-env-audits.js',
   '/js/light-env-evening.js',
   '/js/light-env-model.js',
+  '/js/light-tool-camera-modal-runtime.js',
   '/js/light-tool-camera.js',
   '/js/light-tool-camera-modals.js',
+  '/js/light-tool-cct-meter.js',
+  '/js/light-tool-darkness-meter.js',
+  '/js/light-tool-flicker-detector.js',
+  '/js/light-tool-glass-transmission.js',
+  '/js/light-tool-lux-meter.js',
+  '/js/light-tool-spectrum-classifier.js',
   '/js/light-tools.js',
   '/js/modal-lifecycle.js',
   '/js/marker-analysis.js',
