+++ tests/test-quality-guardrails.js
@@ -204,7 +204,14 @@ const highValueCheckJsModules = [
   'js/lens.js',
   'js/lens-knowledge-base-ui.js',
   'js/light-devices-runtime.js',
+  'js/light-tool-camera-modal-runtime.js',
   'js/light-tool-camera-modals.js',
+  'js/light-tool-cct-meter.js',
+  'js/light-tool-darkness-meter.js',
+  'js/light-tool-flicker-detector.js',
+  'js/light-tool-glass-transmission.js',
+  'js/light-tool-lux-meter.js',
+  'js/light-tool-spectrum-classifier.js',
   'js/pdf-import-commit.js',
   'js/pdf-import.js',
   'js/profile.js',
