+++ tsconfig.checkjs.json
@@ -195,8 +195,15 @@
     "js/light-sun-ai-hooks.js",
     "js/light-sun-loader.js",
     "js/light-today-ai.js",
+    "js/light-tool-camera-modal-runtime.js",
     "js/light-tool-camera.js",
     "js/light-tool-camera-modals.js",
+    "js/light-tool-cct-meter.js",
+    "js/light-tool-darkness-meter.js",
+    "js/light-tool-flicker-detector.js",
+    "js/light-tool-glass-transmission.js",
+    "js/light-tool-lux-meter.js",
+    "js/light-tool-spectrum-classifier.js",
     "js/light-tools-ai-analysis.js",
     "js/light-tools-ui-hooks.js",
     "js/light-tools.js",
